+++ tests/conftest.py
@@ -11,6 +11,10 @@ from pathlib import Path
 import pytest
 import yaml
 
+_REPO_ROOT = Path(__file__).resolve().parents[1]
+if str(_REPO_ROOT) not in sys.path:
+    sys.path.insert(0, str(_REPO_ROOT))
+
 _TEST_DATA_DIR = Path(tempfile.mkdtemp(prefix="ab-tests-"))
 os.environ["AB_DATA_PATH"] = str(_TEST_DATA_DIR)
 _TEST_CONFIG_FILE = _TEST_DATA_DIR / "config.yaml"
@@ -32,19 +36,20 @@ _TEST_CONFIG_FILE.write_text(
     encoding="utf-8",
 )
 
+from anibridge.utils.limiter import Limiter  # noqa: E402
+
 from anibridge.app.config import settings as settings_module  # noqa: E402
 from anibridge.app.config.database import db as db_factory  # noqa: E402
 from anibridge.app.utils import logging as logging_module  # noqa: E402
-from anibridge.app.utils.limiter import Limiter  # noqa: E402
 from anibridge.app.web.state import get_app_state  # noqa: E402
 
 settings_module.get_config.cache_clear()
 logging_module.get_logger.cache_clear()
 db_factory.cache_clear()
 
-src_module = sys.modules.get("src")
+src_module = sys.modules.get("anibridge.app")
 if src_module is None:
-    src_module = importlib.import_module("src")
+    src_module = importlib.import_module("anibridge.app")
 
 src_module.config = settings_module.get_config()  # type: ignore
 src_module.log = logging_module.get_logger()  # type: ignore
